feat: Pull course name from Moodle if unspecified
